refactor: switched to markdown to detail experiences
