+++ mdx-components.tsx
@@ -1,0 +1,6 @@
+import type { MDXComponents } from "mdx/types";
+import { defaultMdxComponents } from "mui-markdown";
+
+export function useMDXComponents(): MDXComponents {
+  return defaultMdxComponents;
+}
