+++ next.config.ts
@@ -1,10 +1,34 @@
 import type { NextConfig } from "next";
+import createMDX from "@next/mdx";
+import remarkGfm from "remark-gfm";
+import remarkMath from "remark-math";
+import rehypeKatex from "rehype-katex";
 
 const nextConfig: NextConfig = {
   /* config options here */
   experimental: {
     typedEnv: true,
   },
+  // Configure `pageExtensions` to include markdown and MDX files
+  pageExtensions: ["js", "jsx", "md", "mdx", "ts", "tsx"],
+  images: {
+    remotePatterns: [
+      {
+        protocol: "https",
+        hostname: "github.com",
+      },
+    ],
+  },
 };
 
-export default nextConfig;
+const withMDX = createMDX({
+  // Handle both .md and .mdx files
+  extension: /\.(md|mdx)$/,
+  options: {
+    remarkPlugins: ["remark-gfm", "remark-math"],
+    rehypePlugins: ["rehype-katex"],
+  },
+});
+
+// Merge MDX config with Next.js config
+export default withMDX(nextConfig);
