+++ src/app/about/page.tsx
@@ -9,17 +9,18 @@ import CardContent from "@mui/material/CardContent";
 import Chip from "@mui/material/Chip";
 import Container from "@mui/material/Container";
 import Grid from "@mui/material/Grid";
-import Stack from "@mui/material/Stack";
+import Stack, { stackClasses } from "@mui/material/Stack";
 import Step from "@mui/material/Step";
 import StepContent from "@mui/material/StepContent";
 import StepLabel from "@mui/material/StepLabel";
 import Typography from "@mui/material/Typography";
 
 import Timeline from "./_components/Timeline";
-import content from "./content";
+import aboutContent from "~/contents/about";
+import Box from "@mui/material/Box";
 
 export default function AboutPage() {
-  const experiences = content.toSorted((a, b) => {
+  const experiences = aboutContent.toSorted((a, b) => {
     if (a.endTime === b.endTime) {
       return a.startTime < b.startTime ? 1 : -1;
     }
@@ -82,122 +83,129 @@ export default function AboutPage() {
             Experience
           </Typography>
           <Timeline orientation="vertical" sx={{ color: "#fff" }}>
-            {experiences.map((exp) => (
-              <Step key={exp.id}>
-                <StepLabel
-                  optional={
-                    <Typography sx={{ color: "#fff" }} variant="caption">
-                      {exp.timePeriod}
-                    </Typography>
-                  }
-                  sx={{
-                    "& .MuiSvgIcon-root": {
-                      color: "#fff !important",
-                      "& circle": {
+            {experiences.map(async (exp) => {
+              const { default: Description } = await import(
+                `~/markdown/${exp.markdown}`
+              );
+              return (
+                <Step key={exp.id}>
+                  <StepLabel
+                    optional={
+                      <Typography sx={{ color: "#fff" }} variant="caption">
+                        {exp.timePeriod}
+                      </Typography>
+                    }
+                    sx={{
+                      "& .MuiSvgIcon-root": {
                         color: "#fff !important",
+                        "& circle": {
+                          color: "#fff !important",
+                        },
+                        "& .MuiStepIcon-text": {
+                          color: "var(--mui-palette-primary-main)",
+                          fill: "var(--mui-palette-primary-main)",
+                          fontWeight: 1000,
+                          fontSize: "0.9rem",
+                        },
                       },
-                      "& .MuiStepIcon-text": {
-                        color: "var(--mui-palette-primary-main)",
-                        fill: "var(--mui-palette-primary-main)",
-                        fontWeight: 1000,
-                        fontSize: "0.9rem",
-                      },
-                    },
-                  }}
-                >
-                  <Typography sx={{ color: "#fff" }}>{exp.position}</Typography>
-                </StepLabel>
-                <StepContent>
-                  <Card>
-                    <CardContent>
-                      <Stack gap={2}>
-                        <Grid columnGap={2} container>
-                          <Grid size={{ xs: 12, sm: 12, md: 12, lg: "auto" }}>
-                            <Typography
-                              color="textSecondary"
-                              fontWeight={700}
-                              variant="h2"
-                            >
-                              {exp.timePeriod}
-                            </Typography>
-                          </Grid>
-                          <Grid size={{ xs: 12, sm: 12, md: 12, lg: "grow" }}>
-                            <Typography fontWeight={700} variant="h2">
-                              {exp.position}
-                            </Typography>
-                          </Grid>
-                          <Grid size={12}>
-                            <Button
-                              endIcon={<ArrowForwardIcon />}
-                              href={exp.company.href}
-                              sx={{
-                                textTransform: "none",
-                                width: "fit-content",
-                                "& svg": {
-                                  fontSize: "2rem !important",
-                                },
-                                "&>span": {
-                                  transition: "100ms ease-in-out",
-                                },
-                                "&:hover > span": {
-                                  ml: "3rem",
-                                },
-                              }}
-                              variant="text"
-                            >
-                              <Typography variant="h4">
-                                {exp.company.name}
+                    }}
+                  >
+                    <Typography sx={{ color: "#fff" }}>
+                      {exp.position}
+                    </Typography>
+                  </StepLabel>
+                  <StepContent>
+                    <Card>
+                      <CardContent>
+                        <Stack gap={2}>
+                          <Grid columnGap={2} container>
+                            <Grid size={{ xs: 12, sm: 12, md: 12, lg: "auto" }}>
+                              <Typography
+                                color="textSecondary"
+                                fontWeight={700}
+                                variant="h2"
+                              >
+                                {exp.timePeriod}
                               </Typography>
-                            </Button>
-                          </Grid>
-                        </Grid>
-                        {exp.description.map((description, index) => (
-                          <Typography key={index}>{description}</Typography>
-                        ))}
-                        {exp.repos && !!exp.repos.length && (
-                          <Stack direction="row" flexWrap="wrap" gap={1}>
-                            {exp.repos.map((link) => (
+                            </Grid>
+                            <Grid size={{ xs: 12, sm: 12, md: 12, lg: "grow" }}>
+                              <Typography fontWeight={700} variant="h2">
+                                {exp.position}
+                              </Typography>
+                            </Grid>
+                            <Grid size={12}>
                               <Button
-                                endIcon={
-                                  <LinkIcon
-                                    sx={{ transform: "rotate(-45deg)" }}
-                                  />
-                                }
-                                href={link.href}
-                                key={link.name}
+                                endIcon={<ArrowForwardIcon />}
+                                href={exp.company.href}
                                 sx={{
+                                  textTransform: "none",
+                                  width: "fit-content",
+                                  "& svg": {
+                                    fontSize: "2rem !important",
+                                  },
                                   "&>span": {
-                                    transition:
-                                      "100ms transform 0s ease-in-out",
+                                    transition: "100ms ease-in-out",
                                   },
                                   "&:hover > span": {
-                                    transform: "rotate(45deg)",
+                                    ml: "3rem",
                                   },
                                 }}
                                 variant="text"
                               >
-                                {link.name}
+                                <Typography variant="h4">
+                                  {exp.company.name}
+                                </Typography>
                               </Button>
-                            ))}
-                          </Stack>
-                        )}
-                        {exp.languages && !!exp.languages.length && (
-                          <Stack direction="row" flexWrap="wrap" gap={1}>
-                            {exp.languages.map((language, index) => (
-                              <Chip
-                                color="primary"
-                                key={index}
-                                label={language}
-                              />
-                            ))}
-                          </Stack>
-                        )}
-                      </Stack>
-                    </CardContent>
-                  </Card>
-                </StepContent>
-              </Step>
-            ))}
+                            </Grid>
+                          </Grid>
+                          <Box sx={{ [`&>.${stackClasses.root}`]: { gap: 2 } }}>
+                            <Description />
+                          </Box>
+                          {exp.repos && !!exp.repos.length && (
+                            <Stack direction="row" flexWrap="wrap" gap={1}>
+                              {exp.repos.map((link) => (
+                                <Button
+                                  endIcon={
+                                    <LinkIcon
+                                      sx={{ transform: "rotate(-45deg)" }}
+                                    />
+                                  }
+                                  href={link.href}
+                                  key={link.name}
+                                  sx={{
+                                    "&>span": {
+                                      transition:
+                                        "100ms transform 0s ease-in-out",
+                                    },
+                                    "&:hover > span": {
+                                      transform: "rotate(45deg)",
+                                    },
+                                  }}
+                                  variant="text"
+                                >
+                                  {link.name}
+                                </Button>
+                              ))}
+                            </Stack>
+                          )}
+                          {exp.languages && !!exp.languages.length && (
+                            <Stack direction="row" flexWrap="wrap" gap={1}>
+                              {exp.languages.map((language, index) => (
+                                <Chip
+                                  color="primary"
+                                  key={index}
+                                  label={language}
+                                />
+                              ))}
+                            </Stack>
+                          )}
+                        </Stack>
+                      </CardContent>
+                    </Card>
+                  </StepContent>
+                </Step>
+              );
+            })}
           </Timeline>
         </Stack>
       </Container>
