+++ src/app/projects/[repo]/_components/ReadMeMarkdown.tsx
@@ -1,17 +1,14 @@
-import "katex/dist/katex.min.css";
+"use server";
 
-import Markdown from "react-markdown";
-import rehypeKatex from "rehype-katex";
-import remarkGfm from "remark-gfm";
-import remarkMath from "remark-math";
 import { getReadme } from "~/api/github";
 import ErrorFallback from "~/components/ErrorFallback";
 import Image from "next/image";
 
-import Stack from "@mui/material/Stack";
-import Typography from "@mui/material/Typography";
+import Stack, { stackClasses } from "@mui/material/Stack";
 import { components } from "@octokit/openapi-types";
-
+import MuiMarkdown from "mui-markdown";
+import Box from "@mui/material/Box";
+import Markdown from "~/components/Markdown";
 interface MarkdownProps {
   owner: string;
   repo: components["schemas"]["full-repository"];
@@ -24,75 +21,14 @@ export default async function ReadMeMarkdown(props: MarkdownProps) {
       .replace(/(?<!!)\[(.*?)\]\((.*?)\)/g, "$1")
       .replace(/<.*>(.*)<\/\w+>/gm, "$1");
 
-    function rewriteImageSrc(src: string) {
-      // If src is a relative path (doesn't start with http)
-      if (!/^https?:\/\//.test(src)) {
-        return `https://github.com/${props.owner}/${props.repo.name}/blob/${props.repo.default_branch}/${src.replace(/^(\.\/|\/)/, "")}?raw=true`;
-      }
-      return src;
-    }
-
     return (
-      <Markdown
-        components={{
-          a: (elementProps) => <Typography>{elementProps.children}</Typography>,
-          p: (elementProps) => <Typography>{elementProps.children}</Typography>,
-          h1: (elementProps) => (
-            <Typography variant="h1">{elementProps.children}</Typography>
-          ),
-          h2: (elementProps) => (
-            <Typography variant="h2">{elementProps.children}</Typography>
-          ),
-          h3: (elementProps) => (
-            <Typography variant="h3">{elementProps.children}</Typography>
-          ),
-          h4: (elementProps) => (
-            <Typography variant="h4">{elementProps.children}</Typography>
-          ),
-          h5: (elementProps) => (
-            <Typography variant="h5">{elementProps.children}</Typography>
-          ),
-          h6: (elementProps) => (
-            <Typography variant="h6">{elementProps.children}</Typography>
-          ),
-          ul: (elementProps) => (
-            <Typography sx={{ pl: 3 }}>
-              <ul>{elementProps.children}</ul>
-            </Typography>
-          ),
-          ol: (elementProps) => (
-            <Typography sx={{ pl: 3 }}>
-              <ol>{elementProps.children}</ol>
-            </Typography>
-          ),
-          table: (elementProps) => (
-            <Typography>{elementProps.children}</Typography>
-          ),
-          img: (elementProps) => {
-            const src = rewriteImageSrc(
-              typeof elementProps.src === "string" ? elementProps.src : ""
-            );
-            return (
-              <Stack
-                alignItems="center"
-                justifyContent="center"
-                sx={{ width: "100%" }}
-              >
-                <Image
-                  alt={elementProps.alt ?? "no alt"}
-                  height={200}
-                  src={src}
-                  width={400}
-                />
-              </Stack>
-            );
-          },
-        }}
-        rehypePlugins={[rehypeKatex]}
-        remarkPlugins={[remarkGfm, remarkMath]}
-      >
-        {readme}
-      </Markdown>
+      <Box sx={{ [`&>.${stackClasses.root}`]: { gap: 2 } }}>
+        <Markdown
+          imageSrc={`https://github.com/${props.owner}/${props.repo.name}/blob/${props.repo.default_branch}/`}
+        >
+          {readme}
+        </Markdown>
+      </Box>
     );
   } catch (error) {
     console.error(error);
