+++ src/contents/about.ts
@@ -1,0 +1,341 @@
+import { Content } from "~/validations/about";
+
+const aboutContent: Content[] = [
+  {
+    id: "1fc63857-1eb6-4779-86bc-2b43872dd00e",
+    startTime: 2024,
+    endTime: 2024,
+    timePeriod: "2024 - Present",
+    type: "work",
+    position: "Software Engineer",
+    company: {
+      name: "Atlantic Technology Ltd.",
+      href: "https://www.atlantictechnology.co.uk/",
+    },
+    repos: [],
+    languages: [
+      "Bash",
+      "C",
+      "C++",
+      "Capacitor",
+      "CMake",
+      "CSS",
+      "Docker",
+      "Electron",
+      "Git",
+      "Go",
+      "gRPC",
+      "HTML",
+      "Java",
+      "JavaScript",
+      "Jenkins",
+      "JNI",
+      "Next.js",
+      "Protobuf",
+      "Python",
+      "Qt",
+      "React",
+      "Rust",
+      "SQLite",
+      "STM32",
+      "Tauri",
+      "TypeScript",
+    ],
+    markdown: "atlantic-technology.md",
+  },
+  {
+    id: "7b05c8e1-e6aa-4a76-bc61-d85d347629be",
+    startTime: 2023,
+    endTime: 2024,
+    timePeriod: "2023 - 2024",
+    type: "work",
+    position: "Lead Software Engineer",
+    company: {
+      name: "Biodevices Without Borders",
+      href: "https://bathbiodevices.com/",
+    },
+    repos: [
+      {
+        name: "Mobile App Code",
+        href: "https://github.com/Bath-Biodevices-Without-Borders/App",
+      },
+      {
+        name: "Micro-controller Code",
+        href: "https://github.com/Bath-Biodevices-Without-Borders/Microcontroller",
+      },
+      {
+        name: "Website Code",
+        href: "https://github.com/Bath-Biodevices-Without-Borders/Website",
+      },
+    ],
+    languages: [
+      "React Native",
+      "React",
+      "Redux",
+      "TypeScript",
+      "JavaScript",
+      "HTML",
+      "CSS",
+      "C++",
+    ],
+    markdown: "biodevices-without-borders.md",
+  },
+  {
+    id: "f5963aa2-c713-4a5e-9789-25904ad8230d",
+    startTime: 2021,
+    endTime: 2022,
+    timePeriod: "2021 - 2022",
+    type: "work",
+    position: "Front-end Web Developer",
+    company: { name: "BT", href: "https://www.bt.com/" },
+    repos: [],
+    languages: [
+      "JavaScript",
+      "HTML",
+      "CSS",
+      "Docker",
+      "Microsoft Power Platform",
+      "Power BI",
+      "Power Apps",
+      "C#",
+      "Unity",
+    ],
+    markdown: "bt.md",
+  },
+  {
+    id: "da17f6e8-1ab8-44de-82db-7dbb51950d9b",
+    startTime: 2020,
+    endTime: 2022,
+    timePeriod: "2020 - 2022",
+    type: "hobby",
+    position: "Game Developer",
+    company: { name: "Independent", href: "/projects" },
+    repos: [
+      {
+        name: "Zombie Runner",
+        href: "https://github.com/George-Madeley/GameDevTV-Unity3D-ZombieRunner",
+      },
+      {
+        name: "Realm Rush",
+        href: "https://github.com/George-Madeley/GameDevTV-Unity3D-RealmRush",
+      },
+      {
+        name: "Obstacle Course",
+        href: "https://github.com/George-Madeley/GameDevTV-Unity3D-ObstacleCourse",
+      },
+      {
+        name: "Project Boost",
+        href: "https://github.com/George-Madeley/GameDevTV-Unity3D-ProjectBoost",
+      },
+      {
+        name: "Argon Assault",
+        href: "https://github.com/George-Madeley/GameDevTV-Unity3D-ArgonAssault",
+      },
+    ],
+    languages: ["C#", "Unity"],
+    markdown: "personal-game-developer.md",
+  },
+  {
+    id: "de1d6887-bc0b-443d-b4e2-cbba06a38723",
+    startTime: 2019,
+    endTime: 2024,
+    timePeriod: "2019 - 2024",
+    type: "education",
+    position: "MEng Computer Systems Engineer",
+    company: { name: "University of Bath", href: "https://www.bath.ac.uk/" },
+    repos: [
+      {
+        name: "AI Sudoku Solver",
+        href: "https://github.com/George-Madeley/CM20252-AI-Sodoku-Solver",
+      },
+      {
+        name: "Circuit Analysis",
+        href: "https://github.com/George-Madeley/EE20084-Circuit-Analysis-Coursework",
+      },
+      {
+        name: "Shared Memory",
+        href: "https://github.com/George-Madeley/CM30225-Shared-Memory-Coursework",
+      },
+      {
+        name: "Distributed Memory",
+        href: "https://github.com/George-Madeley/CM30225-Distributed-Memory-Coursework",
+      },
+      {
+        name: "Image Filtering",
+        href: "https://github.com/George-Madeley/EE40054-Image-Filtering-Coursework",
+      },
+      {
+        name: "Neural Networks",
+        href: "https://github.com/George-Madeley/EE40094-Neural-Networks-Coursework",
+      },
+      {
+        name: "Genetic Algorithms",
+        href: "https://github.com/George-Madeley/EE40094-Genetic-Algorithms-Coursework",
+      },
+      {
+        name: "EEG Spike Classification",
+        href: "https://github.com/George-Madeley/EE40094-EEG-Spike-Classification-Coursework",
+      },
+    ],
+    languages: ["C", "Python", "MATLAB", "Java", "Assembly", "VHDL", "Verilog"],
+    markdown: "university-of-bath.md",
+  },
+  {
+    id: "54642209-cb40-43b0-ab38-e6f956e5e0cb",
+    startTime: 2017,
+    endTime: 2019,
+    timePeriod: "2017 - 2019",
+    type: "education",
+    position: "A-Levels",
+    company: {
+      name: "University Technical College Norfolk",
+      href: "https://www.utcn.org.uk/",
+    },
+    repos: [],
+    languages: [],
+    markdown: "university-technical-college-norfolk.md",
+  },
+  {
+    id: "3fd010c5-c790-478c-bd5e-99dce2e96bd6",
+    startTime: 2018,
+    endTime: 2018,
+    timePeriod: "2018 - 2018",
+    type: "work",
+    position: "Intern",
+    company: {
+      name: "East Coast Energy Internship",
+      href: "https://www.ogdentrust.com/funding/coastal-energy/",
+    },
+    repos: [
+      {
+        name: "Vattenfall",
+        href: "https://www.vattenfall.co.uk/",
+      },
+    ],
+    languages: [],
+    markdown: "east-coast-energy.md",
+  },
+  {
+    id: "c31b9830-d0b0-44e3-806c-5771dfa41a36",
+    startTime: 2017,
+    endTime: 2017,
+    timePeriod: "2017 - 2017",
+    type: "education",
+    position: "Team Manager",
+    company: { name: "F1 in Schools", href: "https://www.f1inschools.co.uk/" },
+    repos: [
+      {
+        name: "Sprowston Community Academy",
+        href: "https://sprowstoncommunityacademy.co.uk/",
+      },
+    ],
+    languages: [],
+    markdown: "sprowston-community-high-school.md",
+  },
+  {
+    id: "96f32e27-bdc2-42e5-8d18-12ada3778892",
+    startTime: 2018,
+    endTime: 2018,
+    timePeriod: "2018 - 2018",
+    type: "work",
+    position: "Researcher",
+    company: {
+      name: "Nuffield Foundation",
+      href: "https://www.nuffieldfoundation.org/",
+    },
+    repos: [
+      {
+        name: "UEA",
+        href: "https://www.uea.ac.uk/",
+      },
+    ],
+    languages: [],
+    markdown: "nuffield-foundation.md",
+  },
+  {
+    id: "580edca2-e2ba-4a74-9e0c-a10b29cbc422",
+    startTime: 2021,
+    endTime: 2022,
+    timePeriod: "2021 - 2022",
+    type: "hobby",
+    position: "Full-Stack Web Developer",
+    company: { name: "Independent", href: "/projects" },
+    repos: [
+      {
+        name: "Past Paper Answer Website",
+        href: "https://github.com/George-Madeley/UoB-Past-Paper-Answer-Website",
+      },
+      {
+        name: "Ravenous",
+        href: "https://github.com/George-Madeley/Ravenous",
+      },
+      {
+        name: "Jamming",
+        href: "https://github.com/George-Madeley/Jammming",
+      },
+      {
+        name: "My Portfolio",
+        href: "https://github.com/George-Madeley/my-portfolio",
+      },
+    ],
+    languages: [
+      "TypeScript",
+      "JavaScript",
+      "React",
+      "Redux",
+      "Express.js",
+      "Node.js",
+      "Vue.js",
+      "JQuery",
+      "HTML",
+      "CSS",
+      "Sass",
+      "SQL",
+    ],
+    markdown: "personal-web.md",
+  },
+  {
+    id: "04c45380-6436-43b1-a383-01850f6e5c1d",
+    startTime: 2023,
+    endTime: 2023,
+    timePeriod: "2023 - 2023",
+    type: "hobby",
+    position: "Machine Learning Engineer",
+    company: { name: "Independent", href: "/projects" },
+    repos: [
+      {
+        name: "Wordle Solver",
+        href: "https://github.com/George-Madeley/AI-Wordle-Solver",
+      },
+      {
+        name: "2048",
+        href: "https://github.com/George-Madeley/AI-2048",
+      },
+      {
+        name: "EEG Spike Classification",
+        href: "https://github.com/George-Madeley/EE40094-EEG-Spike-Classification-Coursework",
+      },
+      {
+        name: "Genetic Algorithms",
+        href: "https://github.com/George-Madeley/EE40094-Genetic-Algorithms-Coursework",
+      },
+      {
+        name: "Neural Networks",
+        href: "https://github.com/George-Madeley/EE40094-Neural-Networks-Coursework",
+      },
+    ],
+    languages: [
+      "Python",
+      "Pandas",
+      "NumPy",
+      "SciPy",
+      "Matplotlib",
+      "Scikit-Learn",
+      "TensorFlow",
+      "Keras",
+    ],
+    markdown: "personal-web.md",
+  },
+];
+
+export default aboutContent;
